CR for client-sdk: moved exceptions mappers from package `mappers` to `exceptions`
