+++ services-gateway-clientsdk/src/main/java/io/scalecube/services/gateway/clientsdk/ClientSettings.java
@@ -1,7 +1,7 @@
 package io.scalecube.services.gateway.clientsdk;
 
-import io.scalecube.services.gateway.clientsdk.exceptions.mappers.ClientErrorMapper;
-import io.scalecube.services.gateway.clientsdk.exceptions.mappers.DefaultClientErrorMapper;
+import io.scalecube.services.gateway.clientsdk.exceptions.ClientErrorMapper;
+import io.scalecube.services.gateway.clientsdk.exceptions.DefaultClientErrorMapper;
 import java.net.InetSocketAddress;
 import reactor.netty.resources.LoopResources;
 import reactor.netty.tcp.SslProvider;
