+++ services-gateway-clientsdk/src/main/java/io/scalecube/services/gateway/clientsdk/exceptions/ClientErrorMapper.java
@@ -1,4 +1,4 @@
-package io.scalecube.services.gateway.clientsdk.exceptions.mappers;
+package io.scalecube.services.gateway.clientsdk.exceptions;
 
 import io.scalecube.services.gateway.clientsdk.ClientMessage;
